TASK fo REVIEW: jd02_03  in progress Note: Issue with Dispithcer class.
