+++ src/by/it/ikavalenka/jd02_03/Dispitcher.java
@@ -2,29 +2,70 @@ package by.it.ikavalenka.jd02_03;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by USER on 31.03.2017.
  */
 public class Dispitcher extends Thread {
-    static final Integer monitorCount = 0;
-     static final int plan = 10;
-    static int countComplete = 0;
-     static int countBuyers = 0;
-    static List<Thread> cashierOne = new ArrayList<>();
+    public static  Integer monitorCount = 0;
+    static final int plan = 10;
+    static AtomicInteger countComplete = new AtomicInteger(0);
+    static AtomicInteger countBuyers = new AtomicInteger(0);
+    static List<Thread> cashiers = new ArrayList<>();
 
+
+    public static void addCountBuyersIn(){
+        try {
+            countComplete.get();
+            countBuyers.get();
+        } finally {
+            cashiers.size();
+        }
+    }
+    public static void addCountBuyersOut(){
+        try {
+            countComplete.get();
+            countBuyers.get();
+        } finally {
+            cashiers.size();
+        }
+    }
+
+    public static void addCountCashier(){
+        try {
+            countComplete.get();
+            countBuyers.get();
+        } finally {
+            cashiers.size();
+        }
+    }
+
+    public static void reduceCountCashier(){
+        try {
+            countComplete.get();
+            countBuyers.get();
+        } finally {
+            cashiers.size();
+        }
+    }
 
     @Override
     public void run() {
-        while (countBuyers < plan) {
-            if (BuyerQueue.getCount() > cashierOne.size() * 5 && cashierOne.size() < 5){
-            Thread cashier = new Thread(new Cashier(cashierOne.size() + 1));
-                cashierOne.add(cashier);
-            cashier.start();
+        ExecutorService executorService= Executors.newFixedThreadPool(5);
+        while (countComplete.get() < plan) {
+            if (BuyerQueue.getCount() > cashiers.size() * 5 && cashiers.size() < 5){
+            Thread cashier = new Thread(new Cashier(cashiers.size() + 1));
+                cashiers.add(cashier);
+//            cashier.start();
+                executorService.execute(cashier);
         }
         Helper.sleep(1000);
 
 //        cashier.remove(this);
+            executorService.shutdown();
     }
         }
 }
