--- src/by/it/ikavalenka/jd02_03/jd02_03/Good.java
@@ -1,27 +1,0 @@
-package by.it.ikavalenka.jd02_03.jd02_03;
-
-class Good {
-    private String name;
-    private double price;
-
-    Good(String name, double price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-}
