--- src/by/it/ikavalenka/jd02_03/jd02_03/IBuyer.java
@@ -1,9 +1,0 @@
-package by.it.ikavalenka.jd02_03.jd02_03;
-
-interface IBuyer {
-    void enterToMarket();
-    void chooseGoods();
-    void goToQueue() throws InterruptedException;
-    void goToOut();
-}
-
