--- src/by/it/ikavalenka/jd02_03/jd02_03/ICashier.java
@@ -1,11 +1,0 @@
-package by.it.ikavalenka.jd02_03.jd02_03;
-
-/**
- * Created by USER on 10.04.2017.
- */
-public interface ICashier {
-    void takeBuyerFromQueue();
-    void accountCashier();
-    void clientMayExit();
-    void calculateProfit();
-}
