--- src/by/it/ikavalenka/jd02_03/jd02_03/IUseBacket.java
@@ -1,7 +1,0 @@
-package by.it.ikavalenka.jd02_03.jd02_03;
-
-interface IUseBacket {
-        void takeBacket();
-        void putGoodsToBucket();
-        void backBacket();
-    }
